backend/requirements.txt
Flask
gunicorn
Flask-Cors
PyYAML
pandas
psycopg[binary]
sympy
pytest
PyJWT>=2.0
cryptography>=41.0.3


